requirements.txt
beautifulsoup4==4.9.3
boto3==1.17.6; (python_version >= "2.7" and python_full_version < "3.0.0") or (python_full_version >= "3.6.0")
botocore==1.20.6; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.6.0"
certifi==2020.12.5; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.5.0"
chardet==4.0.0; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.5.0"
click==7.1.2; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.5.0"
cloudpickle==1.6.0; python_version >= "3.5"
docopt==0.6.2
docutils==0.16; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.5.0"
flask==1.1.2; (python_version >= "2.7" and python_full_version < "3.0.0") or (python_full_version >= "3.5.0")
geographiclib==1.50; python_version >= "3.5"
geopy==2.1.0; python_version >= "3.5"
gunicorn==20.0.4; python_version >= "3.4"
idna==2.10; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.5.0"
itsdangerous==1.1.0; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.5.0"
jinja2==2.11.3; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.5.0"
jmespath==0.10.0; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.6.0"
joblib==1.0.1; python_version >= "3.6"
llvmlite==0.35.0rc3; python_version >= "3.6"
markupsafe==1.1.1; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.5.0"
numba==0.51.2; python_version >= "3.6"
numpy==1.20.1; python_version >= "3.7"
pandas==1.1.5; python_full_version >= "3.6.1"
pymongo==3.11.3
pympler==0.9
python-craigslist==1.1.0
python-dateutil==2.8.1; python_full_version >= "3.6.1"
python-dotenv==0.15.0
pytz==2021.1; python_full_version >= "3.6.1"
requests==2.25.1; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.5.0"
s3transfer==0.3.4; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.6.0"
scikit-learn==0.23.2; python_version >= "3.6"
scipy==1.6.0; python_version >= "3.7"
shap==0.38.1
six==1.15.0; python_full_version >= "3.6.1"
slicer==0.0.7; python_version >= "3.6"
soupsieve==2.2; python_version >= "3.6"
statistics==1.0.3.5
threadpoolctl==2.1.0; python_version >= "3.6"
tqdm==4.56.2; (python_version >= "2.7" and python_full_version < "3.0.0") or (python_full_version >= "3.4.0")
treeinterpreter==0.2.3
urllib3==1.26.3; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version < "4"
werkzeug==1.0.1; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.5.0"
wtforms==2.3.3
